--- modules/MMM-voice control/model/AUTHORS
@@ -1,1 +1,0 @@
-Paul Guyot <pguyot@kallisys.net>
