--- modules/MMM-voice control/model/README.md
@@ -1,91 +1,0 @@
-Modèle Kaldi en français.
-
-Le modèle a été construit avec les scripts Zamia :
-https://github.com/gooofy/zamia-speech
-
-# Licence d'utilisation
-
-Ce modèle est dérivé des corpus textuels et audio listés ci-dessous. Il est publié sous la licence CC BY-NC-SA 4.0 - Creative Common Attribution - Utilisation non commerciale - Partage dans les mêmes conditions 4.0 International.
-
-# Corpus textuels
-
-## Europarl
-
-```
-Europarl: A Parallel Corpus for Statistical Machine Translation, Philipp Koehn, MT Summit 2005
-http://www.statmt.org/europarl/
-```
-
-Licence : non précisée ("not aware of any copyright restrictions")
-
-## Est Républicain
-
-Corpus mis à disposition par le CNRTL
-
-```
-https://cnrtl.fr/corpus/estrepublicain/
-```
-
-Licence : CC BY-NC-SA 2.0
-
-# Corpus audio
-
-## CFPP2000
-
-```
-Branca-Rosoff S., Fleury S., Lefeuvre F., Pires M., 2012, 
-Discours sur la ville. Corpus de Français Parlé Parisien des années 2000 (CFPP2000)
-http://cfpp2000.univ-paris3.fr/CFPP2000.pdf
-```
-
-Licence : CC BY-NC-SA 3.0
-
-## CLAPI
-
-```
-CLAPI, http://clapi.icar.cnrs.fr
-```
-
-Licence : CC BY-NC-SA 4.0
-
-## ESLO-MD
-
-```
-http://eslo.huma-num.fr/
-```
-
-Licence : non précisée ("Accès public")
-
-## M-AILABS Speech Dataset
-
-```
-https://www.caito.de/2019/01/the-m-ailabs-speech-dataset/
-```
-
-Enregistrements par [LibriVox](https://librivox.org/)
-
-Licence : domaine public
-
-## Mozilla Common Voice
-
-````
-https://voice.mozilla.org/fr/datasets
-````
-
-Licence : CC-0
-
-## TCOF
-
-````
-https://www.cnrtl.fr/corpus/tcof/
-````
-
-Licence : CC BY-NC-SA 2.0
-
-## Voxforge
-
-````
-http://www.voxforge.org/
-````
-
-Licence : GPL
