--- modules/MMM-voice control/model/disambig_tid.int
@@ -1,18 +1,0 @@
-16355
-16356
-16357
-16358
-16359
-16360
-16361
-16362
-16363
-16364
-16365
-16366
-16367
-16368
-16369
-16370
-16371
-16372
