--- modules/MMM-voice control/model/ivector/global_cmvn.stats
@@ -1,3 +1,0 @@
- [
-  4.618818e+10 -3.489444e+09 -2.059763e+09 1.921726e+09 -5.841537e+09 -6.048164e+09 -7.75856e+09 -4.873308e+09 -4.763045e+09 -1.330406e+09 -3.983441e+09 -1.278057e+09 -2.862585e+09 -6.033062e+08 -2.847332e+09 -9.149903e+08 -1.936657e+09 -3.699679e+08 -1.09594e+09 -4.810378e+07 -4.374753e+08 4.246159e+07 -7.428497e+07 -1.315271e+07 1.453378e+08 -2.531438e+07 3.035138e+08 -1.289957e+08 2.417447e+08 -2.136708e+08 2.282068e+08 -1.382803e+08 3.409821e+08 7.512533e+07 3.788408e+08 9.763262e+07 1.820446e+08 -1.20177e+08 -1.124585e+08 -1.037246e+08 4.462964e+08 
-  5.022558e+12 2.20806e+11 1.743958e+11 2.220735e+11 2.631863e+11 3.043067e+11 3.330778e+11 2.465113e+11 2.175021e+11 1.655047e+11 1.836205e+11 1.388485e+11 1.366641e+11 1.075725e+11 1.048053e+11 7.125641e+10 6.302351e+10 4.087438e+10 2.970187e+10 1.567274e+10 8.540192e+09 2.760022e+09 3.835788e+08 1.028499e+08 1.296572e+09 3.176209e+09 5.578054e+09 7.290364e+09 8.997982e+09 9.935813e+09 1.04577e+10 1.036794e+10 1.072109e+10 1.081991e+10 9.500429e+09 7.211747e+09 6.830946e+09 5.747362e+09 4.416877e+09 3.020467e+09 0 ]
