--- modules/MMM-voice control/model/ivector/online_cmvn.conf
@@ -1,1 +1,0 @@
-# configuration file for apply-cmvn-online, used in the script ../local/run_online_decoding.sh
