--- modules/MMM-voice control/model/ivector/splice.conf
@@ -1,2 +1,0 @@
---left-context=3
---right-context=3
