--- modules/MMM-voice control/model/mfcc.conf
@@ -1,10 +1,0 @@
-# config for high-resolution MFCC features, intended for neural network training
-# Note: we keep all cepstra, so it has the same info as filterbank features,
-# but MFCC is more easily compressible (because less correlated) which is why 
-# we prefer this method.
---use-energy=false   # use average of log energy, not energy.
---num-mel-bins=40     # similar to Google's setup.
---num-ceps=40     # there is no dimensionality reduction.
---low-freq=20     # low cutoff frequency for mel bins... this is high-bandwidth data, so
-                  # there might be some information at the low end.
---high-freq=-400 # high cutoff frequently, relative to Nyquist of 8000 (=7600)
