--- modules/MMM-voice control/model/word_boundary.int
@@ -1,259 +1,0 @@
-1 nonword
-2 begin
-3 end
-4 internal
-5 singleton
-6 nonword
-7 begin
-8 end
-9 internal
-10 singleton
-11 nonword
-12 begin
-13 end
-14 internal
-15 singleton
-16 begin
-17 end
-18 internal
-19 singleton
-20 begin
-21 end
-22 internal
-23 singleton
-24 begin
-25 end
-26 internal
-27 singleton
-28 begin
-29 end
-30 internal
-31 singleton
-32 begin
-33 end
-34 internal
-35 singleton
-36 begin
-37 end
-38 internal
-39 singleton
-40 begin
-41 end
-42 internal
-43 singleton
-44 begin
-45 end
-46 internal
-47 singleton
-48 begin
-49 end
-50 internal
-51 singleton
-52 begin
-53 end
-54 internal
-55 singleton
-56 begin
-57 end
-58 internal
-59 singleton
-60 begin
-61 end
-62 internal
-63 singleton
-64 begin
-65 end
-66 internal
-67 singleton
-68 begin
-69 end
-70 internal
-71 singleton
-72 begin
-73 end
-74 internal
-75 singleton
-76 begin
-77 end
-78 internal
-79 singleton
-80 begin
-81 end
-82 internal
-83 singleton
-84 begin
-85 end
-86 internal
-87 singleton
-88 begin
-89 end
-90 internal
-91 singleton
-92 begin
-93 end
-94 internal
-95 singleton
-96 begin
-97 end
-98 internal
-99 singleton
-100 begin
-101 end
-102 internal
-103 singleton
-104 begin
-105 end
-106 internal
-107 singleton
-108 begin
-109 end
-110 internal
-111 singleton
-112 begin
-113 end
-114 internal
-115 singleton
-116 begin
-117 end
-118 internal
-119 singleton
-120 begin
-121 end
-122 internal
-123 singleton
-124 begin
-125 end
-126 internal
-127 singleton
-128 begin
-129 end
-130 internal
-131 singleton
-132 begin
-133 end
-134 internal
-135 singleton
-136 begin
-137 end
-138 internal
-139 singleton
-140 begin
-141 end
-142 internal
-143 singleton
-144 begin
-145 end
-146 internal
-147 singleton
-148 begin
-149 end
-150 internal
-151 singleton
-152 begin
-153 end
-154 internal
-155 singleton
-156 begin
-157 end
-158 internal
-159 singleton
-160 begin
-161 end
-162 internal
-163 singleton
-164 begin
-165 end
-166 internal
-167 singleton
-168 begin
-169 end
-170 internal
-171 singleton
-172 begin
-173 end
-174 internal
-175 singleton
-176 begin
-177 end
-178 internal
-179 singleton
-180 begin
-181 end
-182 internal
-183 singleton
-184 begin
-185 end
-186 internal
-187 singleton
-188 begin
-189 end
-190 internal
-191 singleton
-192 begin
-193 end
-194 internal
-195 singleton
-196 begin
-197 end
-198 internal
-199 singleton
-200 begin
-201 end
-202 internal
-203 singleton
-204 begin
-205 end
-206 internal
-207 singleton
-208 begin
-209 end
-210 internal
-211 singleton
-212 begin
-213 end
-214 internal
-215 singleton
-216 begin
-217 end
-218 internal
-219 singleton
-220 begin
-221 end
-222 internal
-223 singleton
-224 begin
-225 end
-226 internal
-227 singleton
-228 begin
-229 end
-230 internal
-231 singleton
-232 begin
-233 end
-234 internal
-235 singleton
-236 begin
-237 end
-238 internal
-239 singleton
-240 begin
-241 end
-242 internal
-243 singleton
-244 begin
-245 end
-246 internal
-247 singleton
-248 begin
-249 end
-250 internal
-251 singleton
-252 begin
-253 end
-254 internal
-255 singleton
-256 begin
-257 end
-258 internal
-259 singleton
